flash: extract TiCI request utilities
